updated component names
